perf-record: replace option --bpf-event with --no-bpf-event
Currently, monitoring of BPF programs through bpf_event is off by default
for perf-record. To turn it on, the user need to use option "--bpf-event".
As BPF gets wider adoption in different subsystems, this option becomes
inconvinient.

This patch makes bpf_event on by default, and adds option "--no-bpf-event"
to turn it off. Since option --bpf-event is not released yet, it is safe
to remove it.

Signed-off-by: Song Liu <songliubraving@fb.com>